day-2022-12-11 part2

+++ day-2022-12-11/Data.cs
@@ -1,14 +1,14 @@
 namespace day_2022_12_11;
 
-public record ItemData(long WorryLevel);
+public record ItemData(int WorryLevel);
 
 public abstract record Operation
 {
-    public record Add(long X) : Operation;
-    public record Mul(long X) : Operation;
+    public record Add(int X) : Operation;
+    public record Mul(int X) : Operation;
     public record Square() : Operation;
 }
 
-public record MonkeyData(IEnumerable<ItemData> Items, Operation Operation, long Divider, int MonkeyTrue, int MonkeyFalse);
+public record MonkeyData(IEnumerable<ItemData> Items, Operation Operation, int Divider, int MonkeyTrue, int MonkeyFalse);
 
 public record Data(IEnumerable<MonkeyData> MonkeyDatas);
